requirements.txt
flask
flask-smorest
python-dotenv
jsonschema
pytest
flask-sqlalchemy
flask-jwt-extended
passlib
